ssr related fixes, subatom.io changes

+++ stores/persistance.ts
@@ -27,6 +27,9 @@ export const localStore = <T extends JsonValue>(key: string, initial: T) => {
 };
 
 export function resetLocalStorage() {
+  if (import.meta.env.SSR) {
+    return;
+  }
   window?.localStorage.clear();
   window?.location.reload();
 }
@@ -35,10 +38,16 @@ export function persistLocally<T extends JsonValue>(
   itemType: ItemType,
   item: T
 ) {
+  if (import.meta.env.SSR) {
+    return;
+  }
   window?.localStorage.setItem(ItemType[itemType], JSON.stringify(item));
 }
 export function retrieveLocally(itemType: ItemType) {
   try {
+    if (import.meta.env.SSR) {
+      return null;
+    }
     let value = window?.localStorage.getItem(ItemType[itemType]);
     if (value) {
       return JSON.parse(value);
